Adds support for W&B logging
Presently the support is somehow hacky: to distinguish between the models
we append the name of the model to the performance measures.
It might be a better way to refactor the script by making the tested model
a parameter of the script, in this way one could launch a sweep using the
tested models as the hyperparameter and W&B would monitor everything we
need (I guess).

It would also be more efficient, we would have also performance measures
(time, mem, etc) of each single method and we could run them in parallel.